Split the traits from the main fractions file
