+++ src/fraction/mod.rs
@@ -1,1 +1,2 @@
+pub mod fraction_traits;
 pub mod fractions;
